fix(make): runs all test categories together
Instead of invoking separate targets it uses + to build list of filters for Ginkgo.

This is later treated as || operation so every test with at least one matching label will be run.

This fixes and issue with coverage report. Now it is accumulated coverage in cover.out file instead of it being overwritten by the subsequent run with different data.