feat(linter): introduces linter to check for problematic code (#41)
This way we don not have to spend time on it during code reviews and can focus on features instead. golangci-lint is set of linters which is highly customizable. This PR enables it both as a make target and GitHub action.

I fixed all the problems so we can start fresh.